Move Repository implementations
This change moves the various repository implementations into their own
private modules and re-exports their types.  No functional changes
expected.
